+++ OggConverter/src/Config/Updates.cs
@@ -35,7 +35,7 @@ namespace OggConverter
         /// WW - week (ex. 18 for 18th week of year)
         /// B - build of this week
         /// </summary>
-        public const int version = 19431;
+        public const int version = 19432;
 
         static bool newUpdateReady;
         static bool newPreviewReady;