+++ OggConverter/src/Forms/Form1.Designer.cs
@@ -313,7 +313,7 @@
             this.btnAbout,
             this.btnQuit});
             this.menuTool.Name = "menuTool";
-            this.menuTool.Size = new System.Drawing.Size(58, 24);
+            this.menuTool.Size = new System.Drawing.Size(58, 26);
             this.menuTool.Text = "Tools";
             // 
             // btnLastLog
@@ -394,14 +394,14 @@
             // menuSettings
             // 
             this.menuSettings.Name = "menuSettings";
-            this.menuSettings.Size = new System.Drawing.Size(76, 24);
+            this.menuSettings.Size = new System.Drawing.Size(76, 26);
             this.menuSettings.Text = "Settings";
             this.menuSettings.Click += new System.EventHandler(this.MenuSettings_Click);
             // 
             // btnLaunchGame
             // 
             this.btnLaunchGame.Name = "btnLaunchGame";
-            this.btnLaunchGame.Size = new System.Drawing.Size(112, 24);
+            this.btnLaunchGame.Size = new System.Drawing.Size(112, 26);
             this.btnLaunchGame.Text = "Launch Game";
             this.btnLaunchGame.Click += new System.EventHandler(this.LaunchTheGameToolStripMenuItem_Click);
             // 
@@ -419,21 +419,21 @@
             this.menu.Location = new System.Drawing.Point(0, 0);
             this.menu.Name = "menu";
             this.menu.RenderMode = System.Windows.Forms.ToolStripRenderMode.System;
-            this.menu.Size = new System.Drawing.Size(991, 28);
+            this.menu.Size = new System.Drawing.Size(991, 30);
             this.menu.TabIndex = 7;
             this.menu.Text = "menu";
             // 
             // btnHelp
             // 
             this.btnHelp.Name = "btnHelp";
-            this.btnHelp.Size = new System.Drawing.Size(55, 24);
+            this.btnHelp.Size = new System.Drawing.Size(55, 26);
             this.btnHelp.Text = "Help";
             this.btnHelp.Click += new System.EventHandler(this.BtnHelp_Click);
             // 
             // btnDonate
             // 
             this.btnDonate.Name = "btnDonate";
-            this.btnDonate.Size = new System.Drawing.Size(122, 24);
+            this.btnDonate.Size = new System.Drawing.Size(122, 26);
             this.btnDonate.Text = "Buy Me a Pizza";
             this.btnDonate.Click += new System.EventHandler(this.btnDonate_Click);
             // 
@@ -816,7 +816,7 @@
             this.btnOpenWithAudacity.Name = "btnOpenWithAudacity";
             this.btnOpenWithAudacity.Size = new System.Drawing.Size(132, 44);
             this.btnOpenWithAudacity.TabIndex = 19;
-            this.btnOpenWithAudacity.Text = "Open with Audacity";
+            this.btnOpenWithAudacity.Text = "Edit with Audacity";
             this.btnOpenWithAudacity.UseVisualStyleBackColor = true;
             this.btnOpenWithAudacity.Click += new System.EventHandler(this.button1_Click);
             //